+++ src/config/ch.sbb.tarifverbundkarte.public/index.js
@@ -4,13 +4,14 @@ import TarifverbundkarteLayer from '../../layers/TarifverbundkarteLayer';
 export const tarifverbundkarteDataLayer = new TrafimageMapboxLayer({
   name: 'ch.sbb.tarifverbundkarte.data',
   visible: true,
-  preserveDrawingBuffer: true,
-  isBaseLayer: true,
-  isQueryable: false,
   zIndex: -1, // Add zIndex as the MapboxLayer would block tiled layers (buslines)
   style: 'ch.sbb.tarifverbund',
   properties: {
     hideInLegend: true,
+    isBaseLayer: true,
+  },
+  mapOptions: {
+    preserveDrawingBuffer: true,
   },
 });
 
@@ -18,6 +19,7 @@ export const tarifverbundkarteLayer = new TarifverbundkarteLayer({
   mapboxLayer: tarifverbundkarteDataLayer,
   visible: true,
   properties: {
+    isQueryable: true,
     hideInLegend: true,
     useOverlay: true,
     popupComponent: 'TarifverbundkartePopup',